+++ lib/ui/settings/settings_drawer.dart
@@ -2665,7 +2665,7 @@ class SettingsSheetState extends State<SettingsSheet> with TickerProviderStateMi
                       AppIcons.coins,
                       onPressed: () async {
                         final String url =
-                            "https://widget.onramper.com?apiKey=${dotenv.env["ONRAMPER_API_KEY"]!}&color=4080D7&onlyCryptos=NANO&defaultCrypto=NANO&&darkMode=true";
+                            "https://widget.onramper.com?apiKey=${dotenv.env["ONRAMPER_API_KEY"]!}&color=4080D7&onlyCryptos=NANO&defaultCrypto=NANO&darkMode=${StateContainer.of(context).curTheme.brightness == Brightness.dark}";
                         await UIUtil.showChromeSafariWebview(context, url);
                       },
                       iconOverride: const SizedBox(